Move to Rock.LibraryLogging namespace/folder
