+++ Rock.Core/LibraryLogging/LibraryLogMessage.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Runtime.CompilerServices;
 
-namespace Rock.Logging.Library
+namespace Rock.LibraryLogging
 {
     /// <summary>
     /// Defines the various properties of a log message.
