+++ Rock.Core/Rock.Core.csproj
@@ -72,12 +72,12 @@
     <Compile Include="Extensions\String\StringExtensions.cs" />
     <Compile Include="Extensions\DateTime\TimeSection.cs" />
     <Compile Include="Extensions\DateTime\ToUnixTimestampExtension.cs" />
-    <Compile Include="Logging\Library\ConsoleLibraryLogger.cs" />
-    <Compile Include="Logging\Library\ILibraryLogger.cs" />
-    <Compile Include="Logging\Library\ILibraryLogMessage.cs" />
-    <Compile Include="Logging\Library\LibraryLogger.cs" />
-    <Compile Include="Logging\Library\LibraryLogMessage.cs" />
-    <Compile Include="Logging\Library\NullLibraryLogger.cs" />
+    <Compile Include="LibraryLogging\ConsoleLibraryLogger.cs" />
+    <Compile Include="LibraryLogging\ILibraryLogger.cs" />
+    <Compile Include="LibraryLogging\ILibraryLogMessage.cs" />
+    <Compile Include="LibraryLogging\LibraryLogger.cs" />
+    <Compile Include="LibraryLogging\LibraryLogMessage.cs" />
+    <Compile Include="LibraryLogging\NullLibraryLogger.cs" />
     <Compile Include="Net\EndpointDetector.cs" />
     <Compile Include="Net\EndpointSelector.cs" />
     <Compile Include="Net\Http\HttpClientFactory.cs" />
